Update Ui by fetching api data from viewmodel

+++ app/src/main/java/com/example/weatherapp/data/model/Main.kt
@@ -5,9 +5,9 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Main (
-    @field:Json(name = "temp") val temp: Float,
-    @field:Json(name = "temp_min") val temp_min: Float,
-    @field:Json(name = "temp_max") val temp_max: Float,
-    @field:Json(name = "pressure") val pressure: Float,
-    @field:Json(name = "humidity") val humidity: Float
+    @field:Json(name = "temp") val temp: Double,
+    @field:Json(name = "temp_min") val temp_min: Double,
+    @field:Json(name = "temp_max") val temp_max: Double,
+    @field:Json(name = "pressure") val pressure: Int,
+    @field:Json(name = "humidity") val humidity: Int
 )
